# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PHP Teste — API de Propostas",
        "description": "API REST para gestão de clientes, propostas financeiras e pedidos. Implementa máquina de estados, optimistic locking, idempotência e auditoria automática.",
        "contact": {
            "name": "Higor Moreira",
            "email": "higor@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/propostas": {
            "get": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Lista propostas com filtros e paginação",
                "operationId": "314cc051691e9bfac937d70db67f723f",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "submitted",
                                "approved",
                                "rejected",
                                "canceled"
                            ]
                        }
                    },
                    {
                        "name": "cliente_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de propostas"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Cria uma nova proposta (status inicial: DRAFT)",
                "operationId": "67a75e44bb7afd71d9e6f32a66e57d12",
                "parameters": [
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "produto",
                                    "valor_mensal",
                                    "origem"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "produto": {
                                        "type": "string",
                                        "example": "Crédito Pessoal"
                                    },
                                    "valor_mensal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1200
                                    },
                                    "origem": {
                                        "type": "string",
                                        "enum": [
                                            "app",
                                            "site",
                                            "api"
                                        ],
                                        "example": "api"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proposta criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropostaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/propostas/{proposta}": {
            "get": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Exibe uma proposta",
                "operationId": "3fd18c6e325de0d3530dc26591e3f716",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da proposta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropostaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Atualiza campos livres com Optimistic Lock",
                "operationId": "fe05489fc32c1e1a39c97968ef95da91",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "versao"
                                ],
                                "properties": {
                                    "versao": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "produto": {
                                        "type": "string",
                                        "example": "Empréstimo"
                                    },
                                    "valor_mensal": {
                                        "type": "number",
                                        "example": 999.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proposta atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropostaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "409": {
                        "description": "Conflito de versão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorConcurrency"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Estado terminal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBusiness"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/propostas/{proposta}/submit": {
            "post": {
                "tags": [
                    "Propostas"
                ],
                "summary": "DRAFT → SUBMITTED",
                "operationId": "f76004f47bcc94df7e4d29761870e247",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposta submetida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropostaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Transição inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBusiness"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/propostas/{proposta}/approve": {
            "post": {
                "tags": [
                    "Propostas"
                ],
                "summary": "SUBMITTED → APPROVED",
                "operationId": "337553da2196ccbec159014ab29067e9",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposta aprovada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropostaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Transição inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBusiness"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/propostas/{proposta}/reject": {
            "post": {
                "tags": [
                    "Propostas"
                ],
                "summary": "SUBMITTED → REJECTED",
                "operationId": "65e051a47abc68aaddfe59ba93a0aa3a",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposta rejeitada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropostaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Transição inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBusiness"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/propostas/{proposta}/cancel": {
            "post": {
                "tags": [
                    "Propostas"
                ],
                "summary": "DRAFT|SUBMITTED → CANCELED",
                "operationId": "911bae092766b19ffcabaa6d23d90e08",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposta cancelada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropostaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Transição inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBusiness"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/propostas/{proposta}/auditoria": {
            "get": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Histórico de auditoria da proposta",
                "operationId": "6514d3da140e6e21e353b43c84f0dec5",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Histórico de auditoria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AuditoriaResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "PropostaResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cliente_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "produto": {
                        "type": "string",
                        "example": "Crédito Pessoal"
                    },
                    "valor_mensal": {
                        "type": "number",
                        "format": "float",
                        "example": 1200
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "submitted",
                            "approved",
                            "rejected",
                            "canceled"
                        ],
                        "example": "draft"
                    },
                    "origem": {
                        "type": "string",
                        "enum": [
                            "app",
                            "site",
                            "api"
                        ],
                        "example": "api"
                    },
                    "versao": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "AuditoriaResource": {
                "properties": {
                    "actor": {
                        "type": "string",
                        "example": "user:1"
                    },
                    "evento": {
                        "type": "string",
                        "enum": [
                            "created",
                            "updated_fields",
                            "status_changed",
                            "deleted_logical"
                        ]
                    },
                    "payload": {
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ErrorBusiness": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unprocessable Entity"
                    },
                    "message": {
                        "type": "string",
                        "example": "Transição de status inválida."
                    }
                },
                "type": "object"
            },
            "ErrorConcurrency": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Conflict"
                    },
                    "message": {
                        "type": "string",
                        "example": "A versão enviada está desatualizada."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Informe o access_token obtido em POST /auth/login",
                "bearerFormat": "OAuth2 (Laravel Passport)",
                "scheme": "bearer"
            }
        }
    },
...
}